+++ tests/acceptance/response_test.py
@@ -234,6 +234,18 @@ def test_app_error_if_path_not_in_spec_and_path_validation_disabled():
         assert app.get('/this/path/doesnt/exist')
 
 
+def test_error_handling_for_12():
+    app = test_app(
+        request=Mock(spec=FixtureRequest, param=['1.2']),
+        **{'pyramid_swagger.enable_response_validation': True}
+    )
+    # Should throw 400 and not 500 (500 is thrown by pyramid_swagger when
+    # response_validation is True and response format does not match the
+    # type specified by the operation's swagger spec. But that match should
+    # be done only when the response status is 200...203)
+    assert app.get('/throw_400', expect_errors=True).status_code == 400
+
+
 def test_response_validation_context():
     request = EnhancedDummyRequest(
         method='GET',